Refactor profile command to better match others
