+++ pctasks/client/setup.cfg
@@ -5,4 +5,4 @@ version = attr: pctasks.client.version.__version__
 pctasks.commands =
     workflow = pctasks.client.workflow.cli:workflow_cmd
     runs = pctasks.client.runs.cli:runs_cmd
-    profile = pctasks.client.profile:profile_cmd
+    profile = pctasks.client.profile.cli:profile_cmd
